User registration add and change logic of access refresh token

+++ src/app/dashboard/dashboard.ts
@@ -6,6 +6,8 @@ import { MatListModule } from '@angular/material/list';
 import { MatIconModule } from '@angular/material/icon';
 import { Router } from '@angular/router';
 import { RouterModule } from '@angular/router';
+import { ConfirmDialog } from '../confirm-dialog/confirm-dialog';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-dashboard',
@@ -18,7 +20,7 @@ export class Dashboard implements OnInit {
   books: BookViewModel[] = [];
   errorMessage: string | null = null;
 
-  constructor(private bookService: Book, private router: Router) { }
+  constructor(private bookService: Book, private router: Router, private dialog: MatDialog) { }
 
   ngOnInit() {
     this.bookService.getAllBooks().subscribe({
@@ -28,19 +30,27 @@ export class Dashboard implements OnInit {
   }
 
   onDeleteBook(id: number): void {
-    this.bookService.deleteBook(id).subscribe({
-      next: (response) => {
-        if (response.success) {
-          this.bookService.getAllBooks().subscribe({
-            next: (data) => this.books = data,
-            error: (err) => this.errorMessage = 'Error loading books after delete'
-          });
-        } else {
-          alert('Delete failed: ' + response.message);
-        }
-      },
-      error: (err) => {
-        alert('Server error: ' + err.message);
+    const dialogRef = this.dialog.open(ConfirmDialog, {
+      width: '250px',
+      data: { message: 'Are you sure you want to delete this book?', btn: 'delete' }
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.bookService.deleteBook(id).subscribe({
+          next: (response) => {
+            if (response.success) {
+              this.bookService.getAllBooks().subscribe({
+                next: (data) => this.books = data,
+                error: (err) => this.errorMessage = 'Error loading books after delete'
+              });
+            } else {
+              alert('Delete failed: ' + response.message);
+            }
+          },
+          error: (err) => {
+            alert('Server error: ' + err.message);
+          }
+        });
       }
     });
   }
